Refactor logic for initializing current user
