+++ frontend/src/features/currentUser/currentUserSlice.js
@@ -1,12 +1,15 @@
-import { createSlice } from '@reduxjs/toolkit'
-import userService from '../../services/user'
+import { createAsyncThunk, createSlice } from '@reduxjs/toolkit'
+import currentUser from '../../services/currentUser'
+
+export const fetchCurrentUser = createAsyncThunk(
+  'currentUser/fetchCurrentUser',
+  currentUser.get,
+)
 
 const userSlice = createSlice({
   initialState: null,
   name: 'currentUser',
   reducers: {
-    set: (state, action) => action.payload,
-    clear: () => null,
     addUpvote: (state, action) => ({
       ...state,
       upvotes: [...state.upvotes, action.payload],
@@ -16,14 +19,21 @@ const userSlice = createSlice({
       upvotes: state.upvotes.filter(id => id !== action.payload),
     }),
   },
+  extraReducers: (builder) => {
+    builder
+      .addCase(fetchCurrentUser.fulfilled, (state, action) => {
+        return {
+          username: action.payload.username,
+          name: action.payload.name,
+          image: action.payload.image,
+          upvotedFeedbacks: [],
+        }
+      })
+      // .addCase()
+      // .addCase()
+  },
 })
 
-export const { set, clear, addUpvote, removeUpvote } = userSlice.actions
-
-export const initializeUser = () => async (dispatch) => {
-  const user = await userService.get()
-  user.upvotes = []
-  dispatch(set(user))
-}
+export const { addUpvote, removeUpvote } = userSlice.actions
 
 export default userSlice.reducer
